requirements.txt
numpy==1.25.1
opencv-python-headless==4.8.0.74
packaging==23.1
Pillow==10.0.0
pytesseract==0.3.10
